requirements.txt
qrcode[pil]==7.4.2
Pillow==10.1.0
requests==2.31.0
flask-socketio==5.3.6
flask==2.3.3
zipstream-ng==1.9.1
pywin32==306
pytest==7.4.3